Closes #8954 - move Hot Rod classes to hotRod package
