+++ model/map-hot-rod/src/main/java/org/keycloak/models/map/storage/hotRod/HotRodMapStorageProvider.java
@@ -19,10 +19,10 @@ package org.keycloak.models.map.storage.hotRod;
 
 import org.keycloak.models.map.common.AbstractEntity;
 import org.keycloak.models.map.common.DeepCloner;
-import org.keycloak.models.map.common.HotRodEntityDescriptor;
+import org.keycloak.models.map.storage.hotRod.common.HotRodEntityDescriptor;
 import org.keycloak.models.map.common.StringKeyConvertor;
 import org.keycloak.models.map.common.UpdatableEntity;
-import org.keycloak.models.map.connections.HotRodConnectionProvider;
+import org.keycloak.models.map.storage.hotRod.connections.HotRodConnectionProvider;
 import org.keycloak.models.map.storage.MapStorage;
 import org.keycloak.models.map.storage.MapStorageProvider;
 import org.keycloak.models.map.storage.MapStorageProviderFactory;
