+++ model/map-hot-rod/src/main/java/org/keycloak/models/map/storage/hotRod/HotRodMapStorageProviderFactory.java
@@ -24,15 +24,15 @@ import org.keycloak.component.AmphibianProviderFactory;
 import org.keycloak.models.ClientModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.KeycloakSessionFactory;
-import org.keycloak.models.map.client.HotRodAttributeEntity;
-import org.keycloak.models.map.client.HotRodClientEntity;
-import org.keycloak.models.map.client.HotRodProtocolMapperEntity;
-import org.keycloak.models.map.common.HotRodPair;
+import org.keycloak.models.map.storage.hotRod.client.HotRodAttributeEntity;
+import org.keycloak.models.map.storage.hotRod.client.HotRodClientEntity;
+import org.keycloak.models.map.storage.hotRod.client.HotRodProtocolMapperEntity;
+import org.keycloak.models.map.storage.hotRod.common.HotRodPair;
 import org.keycloak.models.map.client.MapClientEntity;
 import org.keycloak.models.map.client.MapProtocolMapperEntity;
 import org.keycloak.models.map.common.DeepCloner;
-import org.keycloak.models.map.common.HotRodEntityDescriptor;
-import org.keycloak.models.map.connections.HotRodConnectionProvider;
+import org.keycloak.models.map.storage.hotRod.common.HotRodEntityDescriptor;
+import org.keycloak.models.map.storage.hotRod.connections.HotRodConnectionProvider;
 import org.keycloak.models.map.storage.MapStorageProvider;
 import org.keycloak.models.map.storage.MapStorageProviderFactory;
 import org.keycloak.provider.EnvironmentDependentProviderFactory;
