+++ model/map-hot-rod/src/main/java/org/keycloak/models/map/storage/hotRod/common/HotRodPair.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.models.map.common;
+package org.keycloak.models.map.storage.hotRod.common;
 
 import org.infinispan.protostream.WrappedMessage;
 import org.infinispan.protostream.annotations.ProtoField;
