+++ model/map-hot-rod/src/main/resources/META-INF/services/org.keycloak.models.map.storage.hotRod.connections.HotRodConnectionProviderFactory
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.keycloak.models.map.connections.DefaultHotRodConnectionProviderFactory
+org.keycloak.models.map.storage.hotRod.connections.DefaultHotRodConnectionProviderFactory
